Move useUniqueId
