+++ src/Textarea/Textarea.tsx
@@ -2,11 +2,11 @@ import * as React from "react"
 import CopyToClipboard from "react-copy-to-clipboard"
 
 import Hint from "../Hint/Hint"
-import { useUniqueId } from "../hooks/useUniqueId"
 import Icon from "../Icon/Icon"
 import { LabelText } from "../LabelText/LabelText"
 import { useOperationalContext } from "../OperationalContext/OperationalContext"
 import { DefaultProps } from "../types"
+import { useUniqueId } from "../useUniqueId"
 import { isCmdEnter, lighten } from "../utils"
 import { FormFieldControls, FormFieldError, inputFocus, Label } from "../utils/mixins"
 import styled from "../utils/styled"
